test: rename tests
